Move the existing bulk executor to be provided by a service
