+++ h/services/__init__.py
@@ -94,6 +94,10 @@ def includeme(config):
         ".user_update.user_update_factory", name="user_update"
     )
 
+    config.register_service_factory(
+        ".bulk_executor.BulkExecutorService", name="bulk_executor"
+    )
+
     config.add_directive(
         "add_annotation_link_generator", ".links.add_annotation_link_generator"
     )
