+++ h/views/api/bulk.py
@@ -4,9 +4,6 @@ from logging import getLogger
 from pyramid.response import Response
 
 from h.h_api.bulk_api import BulkAPI
-from h.h_api.bulk_api.executor import AutomaticReportExecutor
-from h.h_api.enums import DataType
-from h.h_api.exceptions import InvalidDeclarationError
 from h.views.api.config import api_config
 
 LOG = getLogger(__name__)
@@ -32,52 +29,6 @@ class FakeBulkAPI(BulkAPI):
             return None
 
 
-class AuthorityCheckingExecutor(AutomaticReportExecutor):
-    """A bulk executor which checks the authority."""
-
-    def __init__(self, authority="lms.hypothes.is"):
-        self.effective_user = None
-        self.authority = authority
-
-    def configure(self, config):
-        self._assert_authority("effective user", config.effective_user)
-
-        self.effective_user = config.effective_user
-
-    def execute_batch(self, command_type, data_type, default_config, batch):
-        for command in batch:
-            self._check_authority(data_type, command.body)
-
-        return super().execute_batch(command_type, data_type, default_config, batch)
-
-    def _assert_authority(self, field, value, embedded=True):
-        if embedded and value.endswith(f"@{self.authority}"):
-            return
-
-        if value == self.authority:
-            return
-
-        raise InvalidDeclarationError(
-            f"The {field} '{value}' does not match the expected authority"
-        )
-
-    def _check_authority(self, data_type, body):
-        if data_type == DataType.USER:
-            self._assert_authority("id", body.id)
-            self._assert_authority(
-                "authority", body.attributes["authority"], embedded=False
-            )
-
-        elif data_type == DataType.GROUP:
-            self._assert_authority("groupid", body.attributes["groupid"])
-            self._assert_authority("query groupid", body.meta["query"]["groupid"])
-
-        elif data_type == DataType.GROUP_MEMBERSHIP:
-            self._assert_authority(
-                "member id", body.relationships["member"]["data"]["id"]
-            )
-
-
 @api_config(
     versions=["v1", "v2"],
     route_name="api.bulk",
@@ -101,9 +52,9 @@ def bulk(request):
     `h.h_api.bulk_api`.
     """
 
-    results = FakeBulkAPI.from_byte_stream(
-        request.body_file, executor=AuthorityCheckingExecutor()
-    )
+    executor = request.find_service(name="bulk_executor").get_executor()
+
+    results = FakeBulkAPI.from_byte_stream(request.body_file, executor=executor)
 
     # No return view is required
     if results is None:
